BP5: complete 32-bit round-trip support and build the 32-bit tests
Finishes the 32<->32 BP5 port on top of the type changes in this branch:

- BP5Writer: the md.idx per-flush (pos,size) records (FlushPosSizeInfo,
  and its gather tmp/RecvBuffer) were size_t, written 4-byte on 32-bit
  but read back as uint64_t -> corrupt flush offsets. Make them uint64_t.
  Also fix MDataTotalSize accounting to sizeof(uint64_t) to match the
  uint64_t header entries it counts. No-op on 64-bit (size_t==uint64_t).
- BPBase: default StatsBlockSize = SIZE_MAX. The former 2^50 truncates to
  0 in a 32-bit size_t, giving a divide-by-zero in BP3/BP4 sub-blocking.
- Tests: cast LocalValueDim/JoinedDim to size_t at Dims braced-init and
  EXPECT_EQ sites (they truncate to size_t on 32-bit, so an un-cast use is
  a -Wnarrowing error in braces and a runtime mismatch in comparisons);
  build LargeBlocks 64-bit only (its RequiredDiskSpace constexpr overflows
  a 32-bit size_t).

+++ testing/adios2/engine/bp/TestBPWriteReadBlockInfo.cpp
@@ -163,8 +163,10 @@ TEST_F(BPWriteReadBlockInfo, BPWriteReadBlockInfo1D8)
         const adios2::Dims start{static_cast<size_t>(Nx * mpiRank)};
         const adios2::Dims count{Nx};
 
-        auto var_local = io.DefineVariable<int32_t>("local", {adios2::LocalValueDim});
-        auto var_localStr = io.DefineVariable<std::string>("localStr", {adios2::LocalValueDim});
+        auto var_local =
+            io.DefineVariable<int32_t>("local", {static_cast<size_t>(adios2::LocalValueDim)});
+        auto var_localStr = io.DefineVariable<std::string>(
+            "localStr", {static_cast<size_t>(adios2::LocalValueDim)});
 
         auto var_iString = io.DefineVariable<std::string>("iString");
         auto var_i8 = io.DefineVariable<int8_t>("i8", shape, start, count);
@@ -760,8 +762,10 @@ TEST_F(BPWriteReadBlockInfo, BPWriteReadBlockInfo1D8_C)
         const adios2::Dims start{static_cast<size_t>(Nx * mpiRank)};
         const adios2::Dims count{Nx};
 
-        auto var_local = io.DefineVariable<int32_t>("local", {adios2::LocalValueDim});
-        auto var_localStr = io.DefineVariable<std::string>("localStr", {adios2::LocalValueDim});
+        auto var_local =
+            io.DefineVariable<int32_t>("local", {static_cast<size_t>(adios2::LocalValueDim)});
+        auto var_localStr = io.DefineVariable<std::string>(
+            "localStr", {static_cast<size_t>(adios2::LocalValueDim)});
 
         auto var_iString = io.DefineVariable<std::string>("iString");
         auto var_i8 = io.DefineVariable<int8_t>("i8", shape, start, count);
